Various fixes for faxes and paper

+++ code/modules/paperwork/paper_bundle.dm
@@ -14,6 +14,8 @@
 	var/amount = 0 //Amount of items clipped to the paper
 	var/page = 1
 	var/screen = 0
+	/// If this was sent via admin fax, allows anyone to see/interact with it
+	var/admin_faxed = FALSE
 
 
 /obj/item/paper_bundle/attackby(obj/item/W as obj, mob/user as mob)
@@ -24,20 +26,20 @@
 		if (istype(P, /obj/item/paper/carbon))
 			var/obj/item/paper/carbon/C = P
 			if (!C.iscopy && !C.copied)
-				user << "<span class='notice'>Take off the carbon copy first.</span>"
+				to_chat(user, span_notice("Take off the carbon copy first."))
 				add_fingerprint(user)
 				return
 		amount++
 		if(screen == 2)
 			screen = 1
-		user << "<span class='notice'>You add [(P.name == "paper") ? "the paper" : P.name] to [(src.name == "paper bundle") ? "the paper bundle" : src.name].</span>"
+		to_chat(user, span_notice("You add [(P.name == "paper") ? "the paper" : P.name] to [(src.name == "paper bundle") ? "the paper bundle" : src.name]."))
 		user.dropItemToGround(P)
 		P.loc = src
 	else if(istype(W, /obj/item/photo))
 		amount++
 		if(screen == 2)
 			screen = 1
-		user << "<span class='notice'>You add [(W.name == "photo") ? "the photo" : W.name] to [(src.name == "paper bundle") ? "the paper bundle" : src.name].</span>"
+		to_chat(user, span_notice("You add [(W.name == "photo") ? "the photo" : W.name] to [(src.name == "paper bundle") ? "the paper bundle" : src.name]."))
 		user.dropItemToGround(W)
 		W.loc = src
 	else if(W.is_hot())
@@ -50,7 +52,7 @@
 			src.amount++
 			if(screen == 2)
 				screen = 1
-		user << "<span class='notice'>You add \the [W.name] to [(src.name == "paper bundle") ? "the paper bundle" : src.name].</span>"
+		to_chat(user, span_notice("You add \the [W.name] to [(src.name == "paper bundle") ? "the paper bundle" : src.name]."))
 		qdel(W)
 	else
 		if(istype(W, /obj/item/pen) || istype(W, /obj/item/toy/crayon))
@@ -86,25 +88,25 @@
 			qdel(src)
 
 		else
-			to_chat(user, "<span class='warning'>You must hold \the [P] steady to burn \the [src].</span>")
+			to_chat(user, span_warning("You must hold \the [P] steady to burn \the [src]."))
 
 /obj/item/paper_bundle/examine(mob/user)
 	if(..(user, 1))
 		src.show_content(user)
 	else
-		user << "<span class='notice'>It is too far away.</span>"
+		to_chat(user, span_notice("It is too far away."))
 	return
 
 /obj/item/paper_bundle/proc/show_content(mob/user as mob)
 	var/dat
 	var/obj/item/W = src[page]
 	dat += "<DIV STYLE='float:left; text-align:left; width:33.33333%'><A href='?src=\ref[src];prev_page=1'>[screen != 0 ? "Previous Page" : ""]</DIV>"
-	dat += "<DIV STYLE='float:left; text-align:center; width:33.33333%'><A href='?src=\ref[src];remove=1'>Remove [(istype(W, /obj/item/paper)) ? "paper" : "photo"]</A></DIV>"
+	dat += "<DIV STYLE='float:left; text-align:center; width:33.33333%'><A href='?src=\ref[src];remove=1'>[admin_faxed ? "" : "Remove [(istype(W, /obj/item/paper)) ? "paper" : "photo"]"]</A></DIV>"
 	dat += "<DIV STYLE='float:left; text-align:right; width:33.33333%'><A href='?src=\ref[src];next_page=1'>[screen != 2 ? "Next Page" : ""]</A></DIV><BR><HR>"
 	if(istype(src[page], /obj/item/paper))
 		var/obj/item/paper/P = W
 		var/dist = get_dist(src, user)
-		if(dist < 2 || istype(usr, /mob/dead/observer) || istype(usr, /mob/living/silicon))
+		if(dist < 2 || admin_faxed)
 			dat += "[P.render_body(user)]<HR>[P.stamps]"
 		else 
 			dat += "[stars(P.render_body(user))]<HR>[P.stamps]"
@@ -136,7 +138,7 @@
 
 /obj/item/paper_bundle/Topic(href, href_list)
 	..()
-	if((src in usr.contents) || (istype(src.loc, /obj/item/folder) && (src.loc in usr.contents)) || IsAdminGhost(usr))
+	if(admin_faxed || (src in usr.contents) || (istype(src.loc, /obj/item/folder) && (src.loc in usr.contents)) || IsAdminGhost(usr))
 		usr.set_machine(src)
 		if(href_list["next_page"])
 			if(page+1 == amount)
@@ -152,33 +154,37 @@
 				return
 			else if(page == 2)
 				screen = 0
-			else if(page == amount+1)
+			else if(page == amount)
 				screen = 1
 			page--
 			playsound(src.loc, "pageturn", 50, 1)
 		if(href_list["remove"])
-			var/obj/item/W = src[page]
-			usr.put_in_hands(W)
-			usr << "<span class='notice'>You remove the [W.name] from the bundle.</span>"
-			if(amount == 1)
+			if(admin_faxed) return // Cannot remove paper from admin faxes
+			amount--
+			if(amount == 0)
 				var/obj/item/paper/P = src[1]
 				usr.dropItemToGround(src)
 				usr.put_in_hands(P)
 				qdel(src)
-			else if(page == amount)
-				screen = 2
-			else if(page == amount+1)
-				page--
+				return
 
-			amount--
+			var/obj/item/W = src[page]
+			usr.put_in_hands(W)
+			to_chat(usr, span_notice("You remove the [W.name] from the bundle."))
+
+			if(page >= amount)
+				page = amount
+			if(page == amount)
+				screen = 2
 			update_icon()
 	else
-		usr << "<span class='notice'>You need to hold it in hand!</span>"
-	if (istype(src.loc, /mob) || istype(src.loc.loc, /mob))
+		to_chat(usr, span_notice("You need to hold it in hand!"))
+	if (istype(src.loc, /mob) || istype(src.loc?.loc, /mob))
 		src.attack_self(src?.loc)
 		updateUsrDialog()
-
-
+	else if (admin_faxed)
+		src.attack_self(usr)
+		updateUsrDialog()
 
 /obj/item/paper_bundle/verb/rename()
 	set name = "Rename bundle"
@@ -197,7 +203,7 @@
 	set category = "Object"
 	set src in usr
 
-	usr << "<span class='notice'>You loosen the bundle.</span>"
+	to_chat(usr, span_notice("You loosen the bundle."))
 	for(var/obj/O in src)
 		O.loc = usr.loc
 		O.layer = initial(O.layer)
@@ -226,7 +232,7 @@
 			underlays += img
 			i++
 		else if(istype(O, /obj/item/photo))
-			var/obj/item/photo/PR
+			var/obj/item/photo/PR = O
 			var/datum/picture/picture2 = PR.picture
 			img = picture2.picture_icon
 			photo = 1
